Sort Functionality
You can now sort when you click on the headers. The headers also change based on what type of sort is applied to that column.

The website now displays a HTML sugfestion when you search for an item that does not exist in the table

There is also CSV Verification, so you can ONLY upload csv's to the program, and no other file types